+++ twilio/rest/messaging/v2/__init__.py
@@ -16,6 +16,7 @@ from typing import Optional
 from twilio.base.version import Version
 from twilio.base.domain import Domain
 from twilio.rest.messaging.v2.channels_sender import ChannelsSenderList
+from twilio.rest.messaging.v2.typing_indicator import TypingIndicatorList
 
 
 class V2(Version):
@@ -28,12 +29,19 @@ class V2(Version):
         """
         super().__init__(domain, "v2")
         self._channels_senders: Optional[ChannelsSenderList] = None
+        self._typing_indicator: Optional[TypingIndicatorList] = None
 
     @property
     def channels_senders(self) -> ChannelsSenderList:
         if self._channels_senders is None:
             self._channels_senders = ChannelsSenderList(self)
         return self._channels_senders
+
+    @property
+    def typing_indicator(self) -> TypingIndicatorList:
+        if self._typing_indicator is None:
+            self._typing_indicator = TypingIndicatorList(self)
+        return self._typing_indicator
 
     def __repr__(self) -> str:
         """